# docs/swagger.yaml
definitions:
...
  model_sys.SysDictType:
    properties:
      dictId:
        description: 字典id
        type: integer
      dictName:
        description: 字典名称
        type: string
      dictType:
        description: 字典类型
        type: string
      status:
        description: 状态
        type: string
    type: object
  model_sys.SysMenu:
    properties:
      component:
        description: 组件路径
        type: string
      icon:
        description: 图标
        type: string
      isFrame:
        description: 是否外链 0 是 1否
        type: string
      menuId:
        description: 菜单id
        type: integer
      menuName:
        description: 菜单名称
        type: string
      menuType:
        description: 菜单类型
        type: string
      orderNum:
        description: 排序
        type: integer
      parentId:
        description: 上级id
        type: integer
      path:
        description: 路由地址
        type: string
      perms:
        description: 权限标识
        type: string
      query:
        description: 参数
        type: string
      status:
        description: 菜单状态（0正常 1停用）
        type: string
      visible:
        description: 菜单状态（0显示 1隐藏）
        type: string
    type: object
  rabbit.RabbitData:
    properties:
      content:
        properties:
          msg:
            type: string
          type:
            type: string
          userId:
            type: string
        type: object
      type:
        type: string
      userId:
        type: string
    type: object
  result.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /demo/rabbit/push:
    post:
      consumes:
      - application/json
      parameters:
      - description: rabbitData
        in: body
        name: rabbitData
        required: true
        schema:
          $ref: '#/definitions/rabbit.RabbitData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: Rabbit发送数据
      tags:
      - Demo - Rabbit
...
  /file/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 上传文件
      tags:
      - Common - 公共接口
  /sys/getLoginUser:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取当前用户的登录信息
      tags:
      - System - 用户管理
  /sys/getRouters:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取前端需要的路由
      tags:
      - System - 菜单管理
  /sys/getUserInfo:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取当前登录用户信息
      tags:
      - System - 用户管理
...
  /sys/sysDict/data/removeCache:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除缓存
      tags:
      - System - 管理
...
  /sys/sysDict/type/{dictId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: dictId
        in: path
        name: dictId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取字典类型详情
      tags:
      - System - 字典管理
  /sys/sysDict/type/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: sysDictType
        in: body
        name: sysDictType
        required: true
        schema:
          $ref: '#/definitions/model_sys.SysDictType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加菜单
      tags:
      - System - 字典管理
  /sys/sysDict/type/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: sysDictType
        in: body
        name: sysDictType
        required: true
        schema:
          $ref: '#/definitions/model_sys.SysDictType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改菜单
      tags:
      - System - 字典管理
  /sys/sysDict/type/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 字典id
        in: query
        name: dictId
        type: integer
      - description: 字典名称
        in: query
        name: dictName
        type: string
      - description: 字典类型
        in: query
        name: dictType
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取字典类型列表
      tags:
      - System - 字典管理
  /sys/sysDict/type/removes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: dictIds
        in: query
        name: dictIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除菜单[逻辑删除]
      tags:
      - System - 管理
  /sys/sysMenu/{menuId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: menuId
        in: path
        name: menuId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取菜单详情
      tags:
      - System - 菜单管理
  /sys/sysMenu/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: sysMenu
        in: body
        name: sysMenu
        required: true
        schema:
          $ref: '#/definitions/model_sys.SysMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加菜单
      tags:
      - System - 菜单管理
  /sys/sysMenu/delete:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: menuId
        in: query
        name: menuId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除菜单[物理删除]
      tags:
      - System - 菜单管理
  /sys/sysMenu/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: sysMenu
        in: body
        name: sysMenu
        required: true
        schema:
          $ref: '#/definitions/model_sys.SysMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改菜单
      tags:
      - System - 菜单管理
  /sys/sysMenu/getTree:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 菜单名
        in: query
        name: menuName
        type: string
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 角色id
        in: query
        name: roleId
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取菜单管理树表
      tags:
      - System - 菜单管理
  /sys/sysMenu/remove:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: menuId
        in: query
        name: menuId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除菜单[逻辑删除]
      tags:
      - System - 菜单管理
swagger: "2.0"